Rename the RobotBuider file to FRC2019DeepSpace to it works with git
so it seems like RobotBuilder name is both the name of the directory it writes to and it ends up in the class path.  so needed it to be putting in the git dirtory so had to have it match the git repository.  After havig code generated I overwrote it with the code from Saturday 2/4/2019